+++ Trunk/Graph/Source/Source/IGraphic.h
@@ -19,7 +19,7 @@ String ReplaceExpression(const String &Str, const TData &Data);
 void Interpolate(std::vector<TPoint> &Points, const std::vector<TPoint> &P, TInterpolationAlgorithm Algorithm);
 double InterpolationAngle(Func32::TDblPoint p1, Func32::TDblPoint p2, Func32::TDblPoint p3, Func32::TDblPoint p4, TInterpolationAlgorithm Algorithm);
 
-TRect Rotate(TMetafile *Metafile, int Degrees);
+TRect Rotate(TMetafile *Metafile, double Degrees);
 
 void InternalGetDIBSizes(HBITMAP Bitmap, int &InfoHeaderSize, int &ImageSize, TPixelFormat PixelFormat);
 bool InternalGetDIB(HBITMAP Bitmap, HPALETTE Palette, void *BitmapInfo, void *Bits, TPixelFormat PixelFormat);
